talks/01_static_islands_dynamic_sea/experiment/requirements.txt
django>=5.1
django-mantle
django-mantle-drf
djangorestframework
drf-spectacular
